requirements.txt
aioconsole
msgspec
aiohttp
pyyaml
click